refactor: move BackAnnotationBuilder to org.manifold.compiler
